UI paint all Cells

+++ Game/vs2012 base project/msvc/BackgroundItem.h
@@ -6,10 +6,10 @@ class BackgroundItem :
     public Item
 {
 public:
-	BackgroundItem(int itemId, float x, float y) :Item(itemId, x, y){}
+	BackgroundItem(int itemId, int x, int y) :Item(itemId, x, y){}
 	void update();
 	void draw(King::Engine& engine);
-	void move(float x, float y);
+	void move(int x, int y);
 	
 	std::tuple<int, int> getBoardLocation();
