+++ Game/vs2012 base project/msvc/JewelryItem.h
@@ -3,10 +3,10 @@
 class JewelryItem :public Item
 {
 public:
-	JewelryItem(int itemId, float x, float y) :Item(itemId,x,y){}
+	JewelryItem(int itemId, int x, int y) :Item(itemId,x,y){}
 	void update();
 	void draw(King::Engine& engine);
-	void move(float x , float y);
+	void move(int x , int y);
 
 };
